LayoutTests/touchadjustment/big-div.html

<html>
<head>
<style>
    #targetDiv {
        background: #00f;
        height: 400px;
        width: 400px;
    }
</style>
<script src="../fast/js/resources/js-test-pre.js"></script>

</head>

<body onload="runTests()">

<div id="targetDiv">
</div>

<p id='description'></p>
<div id='console'></div>

<script>
    var element;
    var adjustedNode;
    var adjustedPoint;
    var targetBounds;
    var touchBounds;

    function findAbsolutePosition(node) {
        var pos = {left: 0, top: 0};
        do {
            pos.left += node.offsetLeft;
            pos.top += node.offsetTop;
        } while (node = node.offsetParent);
        return pos;
    }

    function runTouchTests() {
        element = document.getElementById("targetDiv");
        element.addEventListener('click', function() {}, false);
        document.addEventListener('click', function() {}, false);

        var pos = findAbsolutePosition(element);

        var x = pos.left;
        var y = pos.top;
        var width = element.clientWidth;
        var height = element.clientHeight;

        targetBounds = {
            x: x,
            y: y,
            width: width,
            height: height
        };

        var touchRadius = 10;
        var offset = touchRadius / 2;
        var midX = x + width / 2;
        var midY = y + height / 2;

        debug('\nOverlapping touch above the target should snap to the top of the target element.');
        testTouch(midX, y - offset, touchRadius, targetBounds);
        debug('\nOverlapping touch below the target should snap to the bottom of the target element.');
        testTouch(midX, y + height + offset, touchRadius, targetBounds);
        debug('\nOverlapping touch left of the target should snap to the left side of the target element.');
        testTouch(x - offset, midY, touchRadius, targetBounds);
        debug('\nOverlapping touch right of the target should snap to the right side of the target element.');
        testTouch(x + width + offset, midY, touchRadius, targetBounds);
        debug('\nTest touch area contained within the target element.');
        testTouch(midX, midY, touchRadius, targetBounds);
    }

    function testTouch(touchX, touchY, radius, targetBounds) {

        touchBounds = {
            x: touchX - radius,
            y: touchY - radius,
            width: 2 * radius,
            height: 2 * radius
        };
        adjustedNode = internals.touchNodeAdjustedToBestClickableNode(touchBounds.x, touchBounds.y, touchBounds.width, touchBounds.height, document);
        shouldBe('adjustedNode.id', 'element.id');
        adjustedPoint = internals.touchPositionAdjustedToBestClickableNode(touchBounds.x, touchBounds.y, touchBounds.width, touchBounds.height, document);

        shouldBeTrue('adjustedPoint.x >= targetBounds.x');
        shouldBeTrue('adjustedPoint.x <= targetBounds.x + targetBounds.width');
        shouldBeTrue('adjustedPoint.y >= targetBounds.y');
        shouldBeTrue('adjustedPoint.y <= targetBounds.y + targetBounds.height');
        shouldBeTrue('adjustedPoint.x >= touchBounds.x');
        shouldBeTrue('adjustedPoint.x <= touchBounds.x + touchBounds.width');
        shouldBeTrue('adjustedPoint.y >= touchBounds.y');
        shouldBeTrue('adjustedPoint.y <= touchBounds.y + touchBounds.height');
    }

    function runTests()
    {
        if (window.testRunner && window.internals && internals.touchNodeAdjustedToBestClickableNode) {
            description('Test touch adjustment on a large div.  The adjusted touch point should lie inside the target element and within the touch area.');
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
            runTouchTests();
            isSuccessfullyParsed();
            testRunner.notifyDone();
        }
    }
</script>
</body>
</html>
